fix(roboter-studio): the rest of the block audit, and the verification pass that found what it got wrong
The remainder of the 56-defect block audit, plus a full re-verification of
everything this branch already landed. Splitting further is not honest: after the
lock rework `handlers/motion.py` carries G6, G7, G8 and G13 interleaved, and the
German fixes cross the same test files.

WHAT THE VERIFICATION FOUND, because it is the point of the commit:

Three independent verifiers (claims-by-execution, deep-connections, student
simulation) drove the real Interpreter / WorkflowManager / IKSolver / catalog /
sim over 35 student programs. 26 claims confirmed, 27 of 35 programs better —
and then:

  * THE LOCK FIX WAS ITSELF A REGRESSION. G6 converted the composite motions to a
    re-entering poll loop and left `_run_hat_handler` on a blocking `with`. Mixed
    disciplines on one lock starve the poller outright — measured on a bare
    RLock: blocking waiter 12 acquires, polling waiter 0; uniform polling 11-20.
    A program that finished before now errored, 5/5. `motion.py`'s own comment
    („it is the BOUND, not the asymmetry") was wrong. All EIGHT sites now share
    one discipline; the bound is gone in favour of wait-warn-never-raise, which
    also removes the ten-second budget a breakpoint inside a „wenn …"-Block used
    to give a student before killing the run.
  * A CI GATE WAS RED AND NOTHING COULD SEE IT. A comment reading
    „EDUBOTICS_EDU6_* knobs." made env-forwarding-guard's regex extract
    `EDUBOTICS_EDU6_` — a key that does not exist and is therefore not forwarded.
    No test suite can catch that shape.
  * `MAX_TRAJECTORY_SPAN_S` DID NOT BOUND WHAT IT CLAIMED. It measured the
    endpoints while the cost is per-pair (`round(dt*30)` waypoints per pair), so
    a 114-byte payload with an internal time spike resegmented to 60 002
    waypoints and a 115-byte one to 600 002. Fixed by requiring the time column
    to be non-decreasing, which makes sum(dt) == span and lets the ONE existing
    cap bound the whole stream. Honest recordings are unaffected — the recorder
    stamps monotonic time.
  * WE RE-INTRODUCED, FOUR TIMES, THE DEFECT CLASS THIS ROUND EXISTS TO REMOVE:
    tests written RELATIVE to the constant they pin, which pass at any value
    (1 000 000 for a backlog cap, 1 for an error cap). One test asserted `{}=={}`
    after a field was renamed away and had silently stopped testing anything.
    Every constant now carries a refusal AND an acceptance case with LITERAL
    values, and `test_constant_pins.py` guards the class.
  * GERMAN THAT CONTRADICTED ITS OWN CODE. The reclaim printed „liegt jetzt
    woanders" on the branch literally named `_back_at` — i.e. when the object was
    back exactly where it was — against a docstring three lines up saying the
    sentence states what was observed. Plus gender and case errors on the only
    object in the shipped catalog. Both swept.
  * TWO GERMAN FIXES WOULD HAVE SILENTLY BLINDED LIVE TESTS. A test helper
    filters on the exact sentence being rewritten, after which its assertion
    passes vacuously on the rule it guards. Found by grepping every prescribed
    string against test/ BEFORE prescribing it; both filters fixed in the same
    edit.
  * A GUARD THAT COULD NOT FIRE. `[VAR:name=json]` splits on the FIRST `=`, so a
    variable named with an `=` is truncated and silently overwrites another — and
    the `=` never reaches the name gate we documented as the defence. The
    counters frame written this round is greedy and safe; the two paths
    disagreed.

Also: the reclaim now reasons about POSITION, not absence — an object is done
while it sits where the robot left it, and grabbable again when seen 2 cm away or
back at its pick spot. No absence rule could ever see a student who SLIDES a cube
back (zero absent observations, 0/7 before, 6/7 after), and the previous
two-pass fix made put-back impossible with one cube (0/40). GraspSkip is now
classified by CAUSE: a world problem stays recoverable, a value the student wired
wrongly ends the run — a typo in „schließe um" used to make the arm descend three
times with an open gripper on a green run.

Every constant test mutation-verified: 44/44 killed in this pass, plus 18+20 in
the two implementation groups. Two survivors were real gaps and the tests were
fixed, not the mutations; one was proven an equivalent mutant.

Suites: pytest 1482 -> 1580 passed, 4 skipped. vitest 1071 / 80 files. eslint
clean. GUI 1397, pi_agent 587, cloud 288. compileall 7 roots, compose parity 77
keys, opi twin byte-identical, German lints, enum-parity, tutorials-validate all
pass.

NOT in this commit, deliberately: the VERSION bump (belongs to a release commit),
six items needing owner sign-off (recorded in KNOWN-ISSUES — an unclamped grip
force, a flat-table tail, and three changes that alter when the arm moves), and
any hardware validation. Every measurement here is about the COMMANDED
trajectory; no arm was involved.

+++ physical_ai_tools/physical_ai_server/physical_ai_server/workflow/perception.py
@@ -39,7 +39,22 @@ import numpy as np
 
 @dataclass
 class Detection:
-    centroid_px: tuple[int, int]
+    # Sub-pixel float, for the SAME reason corners_px below is float: this centre
+    # is what handlers/perception_blocks._attach_named_world back-projects through
+    # projection.project_pixel_to_table to get the grasp x/y. It used to be stored
+    # as int(r.center[...]), and int() TRUNCATES rather than rounds, so the error
+    # was a systematic bias, never zero-mean noise. Measured 2026-09-07 against the
+    # real project_pixel_to_table (102 deg HFOV / 640 px -> fx 259 px, camera 0.50 m
+    # above the table, 350 poses): int -> mean dx -0.975 mm, dy +0.927 mm,
+    # mean|err| 1.449 mm, max 2.654 mm; float -> 0.000 mm everywhere. That the bias
+    # is systematic shows in mean|dx| == |mean dx| (0.975 == 0.975) — every pixel
+    # coordinate is positive, so truncation always pulls the same way. The real
+    # detector genuinely carries the information being thrown away: on a rendered
+    # tag36h11 the raw centre came back at .857/.880 px past the integer grid.
+    # The two ROS-wire consumers in physical_ai_server.py already coerce at their
+    # own call site (det.cx = int(cx) for the int32 Detection.msg field, float(cx)
+    # for the verify-point projection), so they are unaffected by construction.
+    centroid_px: tuple[float, float]
     bbox_px: tuple[int, int, int, int]   # x, y, w, h
     confidence: float
     label: str
@@ -162,7 +177,10 @@ class Perception:
         for r in results:
             if aruco_id is not None and r.tag_id != aruco_id:
                 continue
-            cx, cy = int(r.center[0]), int(r.center[1])
+            # Sub-pixel float centre — NOT int(). See Detection.centroid_px for the
+            # measurement: the int cast truncated (never rounded), biasing every
+            # grasp point by ~1.4 mm mean / 2.7 mm max in the same direction.
+            cx, cy = float(r.center[0]), float(r.center[1])
             # Sub-pixel float corners (pupil's native CCW order) for the
             # downstream tag-pose math; a separate int-cast copy drives the
             # bbox so the existing overlay/bbox path is byte-unchanged.
